NU-2436 rename testCases reducer
