+++ designer/client/src/actions/nk/testCasesActions.ts
@@ -1,6 +1,6 @@
 import type { ExpressionObj } from "../../components/graph/node-modal/editors/expression/types";
 import type { TestingDataRecords } from "../../components/modals/TestingDataRecords/Table";
-import type { TestCase } from "../../reducers/graph/testCase";
+import type { TestCases } from "../../reducers/graph/testCases";
 import { getTestCaseAssertions, getTestCaseAssertionsForNode, getInputDataRecords } from "../../reducers/selectors/testCases";
 import type { ThunkAction } from "../reduxTypes";
 
@@ -8,7 +8,7 @@ type Assertion = { expression: ExpressionObj };
 type Assertions = Record<string, Assertion[]>;
 
 export type TestCasesActions =
-    | { type: "TEST_CASE_UPDATE"; testCases: TestCase }
+    | { type: "TEST_CASE_UPDATE"; testCases: TestCases }
     | { type: "SET_TEST_CASE_ASSERTIONS"; assertions: Assertions }
     | { type: "SET_TEST_CASE_INPUTS"; inputs: string };
