+++ designer/client/src/types/scenarioGraph.ts
@@ -1,5 +1,5 @@
 import type { EditorConfig } from "../components/graph/node-modal/editors/expression/EditorConfig";
-import type { TestCase } from "../reducers/graph/testCase";
+import type { TestCases } from "../reducers/graph/testCases";
 import type { ComponentGroup } from "./component";
 import type { TypingResult, UIParameter } from "./definition";
 import type { Edge, EdgeType } from "./edge";
@@ -10,7 +10,7 @@ export type ScenarioGraph = {
     edges: Edge[];
     properties: PropertiesType;
     stickyNotes: NodeType[];
-    testCases: TestCase;
+    testCases: TestCases;
 };
 
 export type Category = string;
